# lib/api-zod/src/generated/types/exchangeRate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UC Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ExchangeRatePackagesItem } from "./exchangeRatePackagesItem";

export interface ExchangeRate {
  rate: number;
  packages: ExchangeRatePackagesItem[];
}